requirements.txt
Flask==2.3.2
Pillow==10.0.0
numpy==1.25.0
werkzeug==2.3.6
gunicorn==20.1.0 